+++ src/nationalarchives/components/error-summary/analytics.js
@@ -1,0 +1,16 @@
+import { valueGetters } from "../../lib/analytics-helpers.mjs";
+
+export default [
+  {
+    scope: ".tna-error-summary",
+    areaName: "error-summary",
+    events: [
+      {
+        eventName: "jump",
+        targetElement: "a.tna-error-summary__link",
+        on: "click",
+        data: { value: valueGetters.text },
+      },
+    ],
+  },
+];