+++ src/nationalarchives/components/text-input/analytics.js
@@ -9,10 +9,9 @@ export default [
         eventName: "blurred",
         on: "blur",
         data: {
-          // eslint-disable-next-line no-unused-vars
           value: valueGetters.value,
           // eslint-disable-next-line no-unused-vars
-          group: ($el, $scope, event) =>
+          group: ($el, $scope) =>
             $scope
               .closest(".tna-form__group")
               ?.querySelector(".tna-form__heading")