+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/nationalarchives/tna-frontend/compare/v0.21.0...HEAD)
 
 ### Added
+
+- New file input component added
+
 ### Changed
 
 - Allow items in an error summary to not have an `href` property